+++ tests/testthat/test-inds-combine.R
@@ -51,8 +51,11 @@ test_that("if multiple names, last kept", {
   expect_identical(wrn$var, c(d = 1L, e = 1L))
 })
 
-test_that("if one name for multiple vars, let caller disambiguate", {
-  expect_equal(inds_combine(letters[1:3], list(x = 1:3)), c(x = 1, x = 2, x = 3))
+test_that("fails if one name for multiple vars", {
+  expect_error(
+    inds_combine(letters[1:3], list(x = 1:3)),
+    "to the same column name"
+  )
 })
 
 test_that("select(0) corner case #82", {